fix: netlify deployment
