--- netlify.toml
@@ -1,11 +1,0 @@
-[build]
-  command   = "npx nx run www:build"
-  publish   = "dist/www/analog/public"
-  functions = "dist/www/analog"
-  ignore    = "npx nx-ignore www"
-
-[context.finance]
-  command   = "npx nx run finance-angular-client:build"
-  publish   = "dist/samples/finance/client/browser"
-  functions = "dist/samples/finance/client/server"
-  ignore    = "npx nx-ignore finance-angular-client"
